Gravity Blast Publishing Touch-Ups
Added code (copied from Ben Fry) that prevents the escape key from closing the application. Commented out the ingame screenshot function. Added a potential app icon. Tidied up the one especially long if statement in the Grid Space object. The main purpose of these changes is to create a version of the sketch that could be exported to an application.